Refactored the App component
